// frontend/FabrikaApp/libs/openapi/src/model/repaymentEntity.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Account } from './account';


export interface RepaymentEntity { 
    id?: number;
    creditorAccount?: Account;
    debtorAccount?: Account;
    amountRepaid?: number;
    repaymentDate?: string;
}